Implement a Discourse-backed blog page
How it works
============

Create the topic in the Newsletter category. If you don't tag it as
`#published`, it won't show up. Once you tag it as published (and only
staff have permission to tag a topic as published), it will show up on
Janitor's blog page.

Also, the webhook payload is not read, and the signature is not
validated. So you can manually trigger a poll by pointing your browser
at that URL.

Importing old entries
=====================

The "slug" in Discourse, which can be seen in the URL, becomes the `#id`
of the post on Janitor (for example, the slug of this topic is
`newsletters-composed-in-discourse-and-published-in-janitor`). If you
don't want to break old links, make sure the post title will result in a
correct slug.

Also, you can click the admin wrench and change the topic timestamp. The
Janitor's blog page sorts posts by their timestamp, so old newsletter
entries will line get sorted correctly.